Adicionados endpoints para geração de extratos. Adicionado Swagger. Classes dtos movidas e projetos renomeados.
