+++ .idea/.idea.TheatricalPlayers/.idea/.name
@@ -1,0 +1,1 @@
+TheatricalPlayers
