--- T2heatricalPlayersRefactoringKata.Tests/StatementPrinterTests.TestStatementExampleLegacy.approved.txt
@@ -1,6 +1,0 @@
-﻿Statement for BigCo
-  Hamlet: $650.00 (55 seats)
-  As You Like It: $547.00 (35 seats)
-  Othello: $456.00 (40 seats)
-Amount owed is $1,653.00
-You earned 47 credits
