--- T2heatricalPlayersRefactoringKata.Tests/StatementPrinterTests.TestTextStatementExample.approved.txt
@@ -1,9 +1,0 @@
-﻿Statement for BigCo
-  Hamlet: $650.00 (55 seats)
-  As You Like It: $547.00 (35 seats)
-  Othello: $456.00 (40 seats)
-  Henry V: $705.40 (20 seats)
-  King John: $931.60 (39 seats)
-  Henry V: $705.40 (20 seats)
-Amount owed is $3,995.40
-You earned 56 credits
