+++ TheatricalPlayers.Core/DataTransferObjects/StatementDTOs/Invoice.cs
@@ -1,0 +1,8 @@
+namespace TheatricalPlayers.Core.DataTransferObjects.StatementDTOs
+{
+    public class Invoice
+    {
+        public string Customer { get; set; }
+        public IReadOnlyList<Performance> Performances { get; set; }
+    }
+}
