+++ TheatricalPlayers.Core/DataTransferObjects/StatementDTOs/Performance.cs
@@ -1,0 +1,8 @@
+namespace TheatricalPlayers.Core.DataTransferObjects.StatementDTOs
+{
+    public class Performance
+    {
+        public int PlayId { get; set; }
+        public int Audience { get; set; }
+    }
+}
