--- TheatricalPlayers.Core/Entities/Invoice.cs
@@ -1,8 +1,0 @@
-namespace TheatricalPlayers.Core.Entities
-{
-    public class Invoice
-    {
-        public string Customer { get; set; }
-        public IReadOnlyList<Performance> Performances { get; set; }
-    }
-}
