--- TheatricalPlayers.Core/Entities/Performance.cs
@@ -1,8 +1,0 @@
-namespace TheatricalPlayers.Core.Entities
-{
-    public class Performance
-    {
-        public int PlayId { get; set; }
-        public int Audience { get; set; }
-    }
-}
