+++ TheatricalPlayers.Core/Interfaces/Statements/IStatementPrinterHandler.cs
@@ -1,4 +1,5 @@
-﻿using TheatricalPlayers.Core.Entities;
+﻿using TheatricalPlayers.Core.DataTransferObjects.StatementDTOs;
+using TheatricalPlayers.Core.Models;
 
 namespace TheatricalPlayers.Core.Interfaces.Statements;
 
@@ -6,4 +7,5 @@ public interface IStatementPrinterHandler
 {
     public string PrintTxt(Invoice invoice, List<Play> plays);
     public string PrintXml(Invoice invoice, List<Play> plays);
+    public Statement GetStatement(Invoice invoice, List<Play> plays);
 }
