+++ TheatricalPlayers.Core/TheatricalPlayers.Core.csproj
@@ -7,4 +7,5 @@
         <Nullable>enable</Nullable>        
     </PropertyGroup>
 
+
 </Project>
