+++ TheatricalPlayers.Tests/Application/Handlers/StatementHandlerTests/PrintTxt.PrintTxtOk.approved.txt
