--- TheatricalPlayersRefactoringKata.Tests/TheatricalPlayersRefactoringKata.Tests.csproj
@@ -1,27 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net8.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-
-        <IsPackable>false</IsPackable>
-        <IsTestProject>true</IsTestProject>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="ApprovalTests" Version="5.7.2" />
-        <PackageReference Include="coverlet.collector" Version="6.0.0"/>
-        <PackageReference Include="FluentAssertions" Version="7.0.0-alpha.4" />
-        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.8.0"/>
-        <PackageReference Include="Moq" Version="4.20.72" />
-        <PackageReference Include="xunit" Version="2.4.1" />
-        
-    </ItemGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\TheatricalPlayers.Application\TheatricalPlayers.Application.csproj" />
-      <ProjectReference Include="..\TheatricalPlayers.Core\TheatricalPlayers.Core.csproj" />
-    </ItemGroup>
-
-</Project>
