--- TheatricalPlayersRefactoringKata/Program.cs
@@ -1,10 +1,0 @@
-﻿namespace TheatricalPlayersRefactoringKata;
-
-public class Program
-{
-    static void Main()
-    {
-        //var test = new StatementPrinter();
-        //test.Print();
-    }
-}
