--- TheatricalPlayersRefactoringKata/TheatricalPlayersRefactoringKata.csproj
@@ -1,14 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net8.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\TheatricalPlayers.Application\TheatricalPlayers.Application.csproj" />
-  </ItemGroup>
-
-</Project>
